script echo fixes, chapter edit code
Added code to address and fix content titles with an ampersand (&) in them. The file names now convert the ampersand to "and" and are correctly retained for metadata within the file without breaking the script or causing everything after the ampersand to drop. As well, full chapter editing code was implemented. If certain chapters are selected out of a PGC, the script now has a function to correct the chapter markers for inclusion into the MKV file. This function can be called from dvd2mkv using a switch on the command line.